+++ pkg/common/versions/installselectors/latest_z_from_delta.go
@@ -44,7 +44,7 @@ func (m latestZFromDelta) SelectVersion(versionList *spi.VersionList) (*semver.V
 			continue
 		}
 
-		if defaultVersion.Version().Patch()+latestZFromDelta == version.Version().Patch() {
+		if int64(defaultVersion.Version().Patch())+latestZFromDelta == int64(version.Version().Patch()) {
 			return version.Version(), versionType, nil
 		}
 	}
